Did refactoring in polymorphic inference
Most of the changes were to clean up the two implementations of type
inference.  As part of this, `Type_constraint` got a new constructor,
`top`, which is mean to be used for typing top-level bindings.

Also, there was a small bug in `Universal_types`, whereby the
quantifiers computed by `Type.get_forall'` were in reverse order with
respect to their order in the type in question.