Used greek letters for certain type operators
The identifier for the record operator is now uppercase pi (`Π`),
instead of `{}`; the variant identifier is now uppercase sigma (`Σ`),
instead of `[]`.  Parsing did not need to change, because there is
special syntax for creating records and variants.  Changing the
identifier for the row kind would have required changes to parsing
(since it does not have a special syntax) so it is still `<>`.