Reworked substitution and beta-reduction
Now, term and type substitution takes the set of free variables as an
argument, and beta-reduction uses the environment to compute the set of
potentially-free variables.  I also fixed a bug in beta-reduction's
application case(s), which incorrectly restricted the environment after
substituting the actual argument into the abstraction body.