Added annotations to `Type_annotation`'s tests
The definitions for `fls` and `tru` now include annotations, so that
they are typed as `A -> A -> A`, instead of `A -> B -> B` and
`A -> B -> B`, respectively.  Similarly, `zero` and `one` are annotated
and typed as `(A -> A) -> A -> A`, instead of `A -> B -> B` and
`(A -> B) -> A -> B`, respectively.